// src/erp-ui/projects/api/src/lib/model/inventory-transaction-type.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum InventoryTransactionType {
    Delivery = 'Delivery',
    Sale = 'Sale',
    Adjustment = 'Adjustment',
    Return = 'Return',
    Loss = 'Loss'
}