// src/erp-ui/projects/api/src/lib/model/stock-overview-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StockOverviewDto { 
    productId: string;
    name: string;
    stock: number;
    threshold: number;
    unit: string;
    purchasePrice: number;
    sellingPriceInternal: number;
    sellingPriceExternal: number;
}