// src/erp-ui/projects/api/src/lib/model/stock-update-request.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StockUpdatePositionDto } from './stock-update-position-dto';


export interface StockUpdateRequest { 
    note: string | null;
    positions: Array<StockUpdatePositionDto>;
}